Move AppHeader2 component to elements dir
