+++ pages/index.vue
@@ -83,7 +83,7 @@ import skills from '~/modules/skills.json'
 import AppBooks from '~/components/AppBooks'
 import AppGifteeButton from '~/components/AppGifteeButton'
 import AppHeader1 from '~/components/elements/AppHeader1'
-import AppHeader2 from '~/components/AppHeader2'
+import AppHeader2 from '~/components/elements/AppHeader2'
 import AppLinkButton from '~/components/elements/AppLinkButton'
 import AppPayPayButton from '~/components/AppPayPayButton'
 import AppProfileIcon from '~/components/AppProfileIcon'
